# Improve Bloomfilter load and save speed by switching to a bit vector