… test/legacyBloomFilter.bin
… test/legacyBloomStrings.txt
… test/speedtest.cpp