Improve Bloomfilter load and save speed by switching to a bit vector which can be read/written directly to disk. Load is 10x faster and save is 100x faster.

<!--
Note: This checklist is a reminder of our shared engineering expectations. Feel free to change it, although assigning a GitHub reviewer and the items in bold are required.
-->

Task/Issue URL: 
Improve binary filter load time #6
https://github.com/duckduckgo/bloom_cpp/issues/6

Tech Design URL: 
CC: 

**Description**:
The approach to speed up load/save time is:

  - Change the vector to a generic bit vector where BlockType is char by default.
  - Add the logic to get/set a bit within a BlockType vector.
  - Add a file header to the saved Bloom Filter that includes version number, magic number, and various size parameters such as maxItems, sizeInBits etc. If the magic number isn't present then assume legacy file layout.
  - To load the vector from a file, get the vector.data() pointer and read directly into that memory address.
  - To write the vector to a file, get the vector.data() pointer and write it out directly.

I did some timing on this new implementation and compared to the previous.
Timing is with 3 million entries in bloomfilter, running on a Macbook Pro.

Lookup time:
new: 2.1 usec
old: 2.0 usec

Insert time:
new: 2.8 usec
old: 2.5 usec

Load time (cumulative all entries):
new: 0.21 sec
old: 2.68 sec

Save time (cumulative):
new: 0.03 sec
old: 2.27 sec
So load is 10x faster, save is 100x faster, lookup and insert 5-10% slower.

**Steps to test this PR**:
1. ./run_test.sh
1. speedtest.cpp to test speed of operations


---
###### Internal references:
[Software Engineering Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design Template](https://app.asana.com/0/59792373528535/184709971311943)