@subsymbolic 2019-11-19
Many thanks for submitting this. Unfortunately we're all on other projects so it may take some time to review this but we will get to it. Thanks again for contributing!