@subsymbolic 2020-09-17
@gdoubleyew, thanks again for your submissions here and apologies for the long silence, we have been on other projects. The good news is we are dusting this repo off and making improvements over the next few days. In terms of this PR, it's a lot bigger than expected, a complete rewrite, well beyond just speeding up the filter.

We can't accept such large-scale changes, however, if you want to package up a minimal footprint of just your data structure and speed changes (no legacy support, headers etc) before we pick this up internally on Monday, that would be great. If not we'll close this out then and continue internally. Thanks again, your ideas and benchmarking will be a huge help to future work either way.