@gdoubleyew 2020-09-18
Ok, thanks for your comments Mia. I probably can't get to this by Monday so I will leave it to your internal process. I can be available for any follow-up questions or actions. Thanks! Grant